navigation-timing/prefetch-transfer-size-executor.html

<!DOCTYPE html>
<html>

<head>
  <title>Navigation Timing Transfert Size of Prefetched Page</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/utils.js"></script>
  <script src="/common/dispatcher/dispatcher.js"></script>
</head>

<body>
  <script>
    const addLink = (url) => {
      return new Promise(resolve => {
        const link = document.createElement('link');
        link.onload = function () { resolve(); };
        link.rel = 'prefetch';
        link.as = 'document';
        link.href = url;
        document.body.appendChild(link);
      });
    }

    const navigateToPrefetchedUrl = (url) => {
      document.location.href = url;
    }

    const getTransferSize = () => window.performance.getEntriesByType('navigation')[0].transferSize;

    promise_test(async t => {
      const testPage = new RemoteContext(token());


      const handler = window.open(`resources/exec.html?uuid=${testPage.context_id}`);
      t.add_cleanup(() => handler.close());

      const prefetchedPage = new RemoteContext(token());

      url_to_prefetch = 'blank_page_prefetch.html?uuid=' + prefetchedPage.context_id;

      // Prefetch link.
      await testPage.execute_script(addLink, [url_to_prefetch]);

      // Navigate to the prefetched link.
      await testPage.execute_script(navigateToPrefetchedUrl, [url_to_prefetch]);

      // Get navigation timing transfer size.
      size = await prefetchedPage.execute_script(getTransferSize);

      assert_equals(0, size);
    }, "Navigation timing transfer size for a prefetched navigation should be 0.");
  </script>
</body>

</html>
